app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asynch==0.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
clickhouse-driver==0.2.7
clickhouse-sqlalchemy==0.3.1
fastapi==0.110.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
isort==5.13.2
leb128==1.0.7
lz4==4.3.3
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
zstd==1.5.5.1
